+++ app/src/main/java/com/linari/data/checking/models/CheckingListGroupedModel.kt
@@ -18,7 +18,7 @@ data class CheckingListGroupedRow(
     @SerializedName("CustomerCode")
     val customerCode: String,
     @SerializedName("CustomerID")
-    val customerID: Int,
+    val customerID: Long,
     @SerializedName("CustomerName")
     val customerName: String,
     @SerializedName("CustomerTypeTitle")
