+++ app/src/main/java/com/linari/data/checking/models/PalletManifestInfo.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 
 data class PalletManifestInfo(
     @SerializedName("HasPallet") val hasPallet: Boolean,
-    @SerializedName("PalletStatusID") val palletStatusID: Int?,
+    @SerializedName("PalletStatusID") val palletStatusID: Long?,
     @SerializedName("PalletStatusTitle")val palletStatusTitle: String?,
-    @SerializedName("PalletTypeID") val palletTypeID: Int?,
+    @SerializedName("PalletTypeID") val palletTypeID: Long?,
     @SerializedName("PalletTypeTitle") val palletTypeTitle: String?
 )
